refactor: move navbar to separate folder
